requirements.txt
playwright
openai
pydantic
beautifulsoup4
python-dotenv
